requirement.txt
streamlit
yfinance
pandas
plotly
pandas-ta-classic
numpy
mplfinance